Update logrosActions to get better performance with listeners

+++ src/actions/logrosActions.js
@@ -3,7 +3,7 @@ import { LOAD_USER_VERIFICATION_STATUS, LOAD_LOGROS_ACTIVOS, REMOVE_LOGRO_ACTIVO
 
 export const loadQaplaLogros = (uid) => async (dispatch) => {
     cuentasVerificadasRef.child(uid).on('value', (verifiedAccount) => {
-        if (verifiedAccount.exists()) {
+        if (verifiedAccount.exists() && verifiedAccount.val().status === 2) {
             dispatch(checkIfIsUserVerifiedSuccess(true));
             dispatch(removeLogroFromActivos('Verification-Logro'));
         } else {
@@ -12,10 +12,10 @@ export const loadQaplaLogros = (uid) => async (dispatch) => {
         }
     });
     
-    logrosActRef.on('value', (logrosActivos) => {
-        logrosActivos.forEach((logroActivoType) => {
-            if (logroActivoType.key !== 'verifica') {
-                logroActivoType.forEach((logroActivo) => {
+    logrosActRef.once('value', (logrosActivos) => {
+        logrosActivos.forEach((typeOfLogro) => {
+            if (typeOfLogro.key !== 'verifica') {
+                typeOfLogro.forEach((logroActivo) => {
                     const logroActivoObject = {
                         id: logroActivo.key,
                         ...logroActivo.val()
@@ -24,31 +24,32 @@ export const loadQaplaLogros = (uid) => async (dispatch) => {
                 });
             }
         });
-        logrosRef.child(uid).child('logroIncompleto').on('value', (logrosIncompletos) => {
-            if (logrosIncompletos.exists()) {
-                logrosIncompletos.forEach((logroIncompleto) => {
-                    const logroIncompletoObject = {
-                        id: logroIncompleto.key,
-                        ...logroIncompleto.val()
-                    };
-                    dispatch(loadLogrosActivosSuccess(logroIncompletoObject));
-                });
-            }
-        });
+    });
 
-        logrosRef.child(uid).child('logroCompleto').on('value', (logrosCompletos) => {
-            if (logrosCompletos.exists()) {
-                logrosCompletos.forEach((logroCompleto) => {
-                    const logroCompletoObject = {
-                        id: logroCompleto.key,
-                        tipoLogro: 'completado',
-                        ...logroCompleto.val()
-                    }
-                    dispatch(loadLogrosCompletosSuccess(logroCompletoObject));
-                    dispatch(removeLogroFromActivos(logroCompleto.key));
-                });
-            }
-        });
+    logrosRef.child(uid).child('logroIncompleto').on('value', (logrosIncompletos) => {
+        if (logrosIncompletos.exists()) {
+            logrosIncompletos.forEach((logroIncompleto) => {
+                const logroIncompletoObject = {
+                    id: logroIncompleto.key,
+                    ...logroIncompleto.val()
+                };
+                dispatch(loadLogrosActivosSuccess(logroIncompletoObject));
+            });
+        }
+    });
+
+    logrosRef.child(uid).child('logroCompleto').on('child_added', (logroCompleto) => {
+        const logroCompletoObject = {
+            id: logroCompleto.key,
+            tipoLogro: 'completado',
+            ...logroCompleto.val()
+        }
+        dispatch(loadLogrosCompletosSuccess(logroCompletoObject));
+        dispatch(removeLogroFromActivos(logroCompleto.key));
+    });
+
+    logrosActRef.on('child_removed', (removedLogro) => {
+        dispatch(removeLogroFromActivos(removedLogro.key));
     });
 }
